Revert "KVM: x86: preserve the high 32-bits of the PAT register"
euler inclusion
category: kabi

Avoid KABI breakage. We need to revisit this later.

------------------------

This reverts commit 7cb060a91c0efc5ff94f83c6df3ed705e143cdb9.

+++ arch/x86/include/asm/kvm_host.h
@@ -437,7 +437,7 @@ struct kvm_vcpu_arch {
 	bool nmi_injected;    /* Trying to inject an NMI this entry */
 
 	struct mtrr_state_type mtrr_state;
-	u64 pat;
+	u32 pat;
 
 	int switch_db_regs;
 	unsigned long db[KVM_NR_DB_REGS];
